+++ routes/breadcrumbs.php
@@ -72,3 +72,23 @@ Breadcrumbs::register('newsletter-edit', function ($breadcrumbs): void {
     $breadcrumbs->push('Nieuwsbrieven', route('admin.nieuwsbrief.index')); 
     $breadcrumbs->push('Wijzig nieuwsbrief', route('admin.nieuwsbrief.index'));
 });
+
+Breadcrumbs::register('categories-index', function ($breadcrumbs): void {
+    $breadcrumbs->push('Home', url('home')); 
+    $breadcrumbs->push('Nieuws', route('admin.articles.index'));
+    $breadcrumbs->push('Categorieen', route('admin.categories.index'));
+});
+
+Breadcrumbs::register('categories-create', function ($breadcrumbs): void {
+    $breadcrumbs->push('Home', url('home')); 
+    $breadcrumbs->push('Nieuws', route('admin.articles.index'));
+    $breadcrumbs->push('Categorieen', route('admin.categories.index'));
+    $breadcrumbs->push('Nieuw', route('admin.categories.create'));
+});
+
+Breadcrumbs::register('categories-edit', function ($breadcrumbs): void {
+    $breadcrumbs->push('Home', url('home')); 
+    $breadcrumbs->push('Nieuws', route('admin.articles.index'));
+    $breadcrumbs->push('Categorieen', route('admin.categories.index'));
+    $breadcrumbs->push('Wijzig', route('admin.categories.create'));
+});