+++ CHANGELOG.md
@@ -7,4 +7,7 @@ All notable changes to `misfits-BE/activisme.be` will be documented in this file
 - Fix #64 | Implementatie changelog
 - Fix #71 | Global user variable.
 - Fix #73 | Datum notitie in formulier is verkeerd
-- Fix #74 | Gebruiker moet kalender evenement kunnen wijzigen. 
+- Fix #75 | Implementatie `nunomaduro/collision` package.
+- Fix #77 | Beheers console voor de categories v/d nieuwsberichten. 
+- Fix #78 | Implementatie dropdown voor nieuws categorieen. 
+- Fix #74 | Gebruiker moet kalender evenement kunnen wijzigen.